Move property classes to postconf_master subclass
Custom properties can cause issues with autoloading the type because
the ruby classes are not available on the master load path. Since they
are used exclusively in postconf_master anyway, moving them there avoids
those issues (compare puppet's cron type).
